# dubaimedia/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dubai Media Gateway API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: session
    description: Authentication / session
  - name: eligibility
    description: Rebate eligibility checker
  - name: feedback
    description: AI pitch feedback
  - name: membership
    description: Membership tiers
  - name: talent
    description: National talent registry
  - name: crewMatch
    description: Crew matching engine
  - name: venues
    description: Venues catalog and bookings
  - name: hub
    description: Community hub applications
  - name: investors
    description: Investor matching
  - name: festival
    description: Film festival submissions
  - name: dashboard
    description: User dashboard
  - name: partner
    description: Partner KPI dashboard

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ─── Session ────────────────────────────────────────────────────────────────
  /session:
    get:
      operationId: getSession
      tags: [session]
      summary: Get current session user
      responses:
        "200":
          description: Current user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
    post:
      operationId: createSession
      tags: [session]
      summary: Sign in with email (simulated auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
    delete:
      operationId: deleteSession
      tags: [session]
      summary: Sign out
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Membership ─────────────────────────────────────────────────────────────
  /membership:
    post:
      operationId: joinMembership
      tags: [membership]
      summary: Upgrade membership tier (simulated checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInput"
      responses:
        "200":
          description: Membership updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"

  # ─── Eligibility ─────────────────────────────────────────────────────────────
  /eligibility:
    get:
      operationId: listEligibilityApplications
      tags: [eligibility]
      summary: List current user's eligibility applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EligibilityApplication"
    post:
      operationId: submitEligibility
      tags: [eligibility]
      summary: Submit eligibility check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityInput"
      responses:
        "201":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityApplication"

  # ─── Feedback ─────────────────────────────────────────────────────────────
  /feedback:
    get:
      operationId: listFeedbackAnalyses
      tags: [feedback]
      summary: List current user's pitch analyses
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PitchAnalysis"
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit pitch for AI feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PitchAnalysis"
        "402":
          description: Free analysis limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaywallResponse"

  # ─── Talent ─────────────────────────────────────────────────────────────────
  /talent:
    get:
      operationId: listTalent
      tags: [talent]
      summary: Browse talent registry
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: emiratiOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of talent profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TalentProfile"
    post:
      operationId: registerTalent
      tags: [talent]
      summary: Register in talent directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TalentInput"
      responses:
        "201":
          description: Registration submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TalentProfile"

  # ─── Crew Match ──────────────────────────────────────────────────────────────
  /crew-match:
    post:
      operationId: matchCrew
      tags: [crewMatch]
      summary: Match crew for a production
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewMatchInput"
      responses:
        "200":
          description: Proposed crew team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMatchResult"

  /crew-match/request:
    post:
      operationId: requestCrewTeam
      tags: [crewMatch]
      summary: Save a crew team request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewRequestInput"
      responses:
        "201":
          description: Team request saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Venues ─────────────────────────────────────────────────────────────────
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: Browse venues catalog
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"

  /venues/{id}/booking:
    post:
      operationId: requestVenueBooking
      tags: [venues]
      summary: Request a venue booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Hub ─────────────────────────────────────────────────────────────────────
  /hub:
    post:
      operationId: applyForHub
      tags: [hub]
      summary: Submit hub workspace application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HubApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Investors ───────────────────────────────────────────────────────────────
  /investors:
    get:
      operationId: listInvestorProjects
      tags: [investors]
      summary: List investor project cards (Studio tier required)
      responses:
        "200":
          description: Investor projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestorProject"
        "403":
          description: Studio tier required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaywallResponse"

  /investors/{id}/interest:
    post:
      operationId: expressInvestorInterest
      tags: [investors]
      summary: Express interest in an investor project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interest logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Festival ────────────────────────────────────────────────────────────────
  /festival:
    post:
      operationId: submitFestivalEntry
      tags: [festival]
      summary: Submit a festival entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FestivalSubmissionInput"
      responses:
        "201":
          description: Submission confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FestivalSubmission"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get all user applications and submissions
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"

  # ─── Partner ─────────────────────────────────────────────────────────────────
  /partner:
    get:
      operationId: getPartnerKpis
      tags: [partner]
      summary: Get partner KPI dashboard data
      responses:
        "200":
          description: KPI data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerKpis"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    PaywallResponse:
      type: object
      properties:
        error:
          type: string
        upgradeRequired:
          type: string
      required:
        - error
        - upgradeRequired

    # ─── Session ──────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        tier:
          type: string
          enum: [free, creator, studio]
        analysisCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - email
        - tier
        - analysisCount
        - createdAt

    SessionResponse:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required:
        - user

    SessionInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
      required:
        - email

    MembershipInput:
      type: object
      properties:
        tier:
          type: string
          enum: [creator, studio]
        name:
          type: string
        email:
          type: string
        cardNumber:
          type: string
      required:
        - tier
        - name
        - email

    # ─── Eligibility ──────────────────────────────────────────────────────────
    EligibilityInput:
      type: object
      properties:
        projectType:
          type: string
        projectTitle:
          type: string
        logline:
          type: string
        budgetAed:
          type: number
        dubaiSpendPercent:
          type: number
        shootDays:
          type: integer
        emiratiCrewPercent:
          type: number
        usesDubaiStudioCity:
          type: boolean
        featuresDubaiLocations:
          type: boolean
        scriptFilename:
          type: ["string", "null"]
        scriptFileSize:
          type: ["integer", "null"]
      required:
        - projectType
        - projectTitle
        - logline
        - budgetAed
        - dubaiSpendPercent
        - shootDays
        - emiratiCrewPercent
        - usesDubaiStudioCity
        - featuresDubaiLocations

    RebateBreakdown:
      type: object
      properties:
        baseRate:
          type: number
        emiratiBonus:
          type: number
        studioBonus:
          type: number
        locationBonus:
          type: number
        totalRate:
          type: number
        estimatedRebateAed:
          type: number
        qualifyingSpendAed:
          type: number
        meetsMinimum:
          type: boolean
      required:
        - baseRate
        - emiratiBonus
        - studioBonus
        - locationBonus
        - totalRate
        - estimatedRebateAed
        - qualifyingSpendAed
        - meetsMinimum

    EligibilityApplication:
      type: object
      properties:
        id:
          type: string
        userId:
          type: ["string", "null"]
        projectType:
          type: string
        projectTitle:
          type: string
        logline:
          type: string
        budgetAed:
          type: number
        dubaiSpendPercent:
          type: number
        shootDays:
          type: integer
        emiratiCrewPercent:
          type: number
        usesDubaiStudioCity:
          type: boolean
        featuresDubaiLocations:
          type: boolean
        scriptFilename:
          type: ["string", "null"]
        rebate:
          $ref: "#/components/schemas/RebateBreakdown"
        status:
          type: string
        submittedAt:
          type: string
      required:
        - id
        - projectType
        - projectTitle
        - logline
        - budgetAed
        - dubaiSpendPercent
        - shootDays
        - emiratiCrewPercent
        - usesDubaiStudioCity
        - featuresDubaiLocations
        - rebate
        - status
        - submittedAt

    # ─── Feedback ─────────────────────────────────────────────────────────────
    FeedbackInput:
      type: object
      properties:
        logline:
          type: string
        synopsis:
          type: string
        genre:
          type: string
        targetAudience:
          type: string
        format:
          type: string
      required:
        - logline
        - genre
        - targetAudience
        - format

    PitchScore:
      type: object
      properties:
        dimension:
          type: string
        score:
          type: integer
      required:
        - dimension
        - score

    PitchAnalysis:
      type: object
      properties:
        id:
          type: string
        userId:
          type: ["string", "null"]
        logline:
          type: string
        genre:
          type: string
        targetAudience:
          type: string
        format:
          type: string
        scores:
          type: array
          items:
            $ref: "#/components/schemas/PitchScore"
        strengths:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        nextSteps:
          type: array
          items:
            type: string
        submittedAt:
          type: string
      required:
        - id
        - logline
        - genre
        - targetAudience
        - format
        - scores
        - strengths
        - risks
        - nextSteps
        - submittedAt

    # ─── Talent ───────────────────────────────────────────────────────────────
    TalentInput:
      type: object
      properties:
        name:
          type: string
        nationality:
          type: string
        department:
          type: string
        yearsExperience:
          type: integer
        portfolioUrl:
          type: string
        bio:
          type: string
      required:
        - name
        - nationality
        - department
        - yearsExperience

    TalentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nationality:
          type: string
        department:
          type: string
        yearsExperience:
          type: integer
        rating:
          type: number
        isVerified:
          type: boolean
        isEmirati:
          type: boolean
        portfolioUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending]
      required:
        - id
        - name
        - nationality
        - department
        - yearsExperience
        - rating
        - isVerified
        - isEmirati
        - status

    # ─── Crew Match ───────────────────────────────────────────────────────────
    CrewMatchInput:
      type: object
      properties:
        projectType:
          type: string
        departments:
          type: array
          items:
            type: string
        shootWindow:
          type: string
      required:
        - projectType
        - departments
        - shootWindow

    CrewMember:
      type: object
      properties:
        department:
          type: string
        profile:
          oneOf:
            - $ref: "#/components/schemas/TalentProfile"
            - type: "null"
        isPlaceholder:
          type: boolean
        placeholderLabel:
          type: ["string", "null"]
      required:
        - department
        - isPlaceholder

    CrewMatchResult:
      type: object
      properties:
        projectType:
          type: string
        shootWindow:
          type: string
        team:
          type: array
          items:
            $ref: "#/components/schemas/CrewMember"
      required:
        - projectType
        - shootWindow
        - team

    CrewRequestInput:
      type: object
      properties:
        projectType:
          type: string
        departments:
          type: array
          items:
            type: string
        shootWindow:
          type: string
      required:
        - projectType
        - departments
        - shootWindow

    # ─── Venues ───────────────────────────────────────────────────────────────
    Venue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        location:
          type: string
        capacity:
          type: integer
        dayRateAed:
          type: integer
        tags:
          type: array
          items:
            type: string
        available:
          type: boolean
        description:
          type: string
      required:
        - id
        - name
        - type
        - location
        - capacity
        - dayRateAed
        - tags
        - available
        - description

    BookingInput:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        productionTitle:
          type: string
        contactEmail:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - startDate
        - endDate
        - productionTitle
        - contactEmail

    # ─── Hub ─────────────────────────────────────────────────────────────────
    HubApplicationInput:
      type: object
      properties:
        workspaceType:
          type: string
          enum: [hot_desk, resident_desk, edit_suite, screening_room]
        startDate:
          type: string
        teamSize:
          type: integer
        projectSummary:
          type: string
        contactEmail:
          type: string
      required:
        - workspaceType
        - startDate
        - teamSize
        - projectSummary
        - contactEmail

    # ─── Investors ────────────────────────────────────────────────────────────
    InvestorProject:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        genre:
          type: string
        budgetAed:
          type: integer
        seekingAed:
          type: integer
        rebatePrequalified:
          type: boolean
        synopsis:
          type: string
        interestExpressed:
          type: boolean
      required:
        - id
        - title
        - genre
        - budgetAed
        - seekingAed
        - rebatePrequalified
        - synopsis
        - interestExpressed

    # ─── Festival ─────────────────────────────────────────────────────────────
    FestivalSubmissionInput:
      type: object
      properties:
        filmTitle:
          type: string
        runtime:
          type: integer
        category:
          type: string
        premiereStatus:
          type: string
        screenerUrl:
          type: string
        directorName:
          type: string
        contactEmail:
          type: string
      required:
        - filmTitle
        - runtime
        - category
        - premiereStatus
        - screenerUrl
        - directorName
        - contactEmail

    FestivalSubmission:
      type: object
      properties:
        id:
          type: string
        submissionId:
          type: string
        filmTitle:
          type: string
        runtime:
          type: integer
        category:
          type: string
        premiereStatus:
          type: string
        directorName:
          type: string
        status:
          type: string
        submittedAt:
          type: string
      required:
        - id
        - submissionId
        - filmTitle
        - runtime
        - category
        - premiereStatus
        - directorName
        - status
        - submittedAt

    # ─── Dashboard ────────────────────────────────────────────────────────────
    DashboardItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        status:
          type: string
        createdAt:
          type: string
        detail:
          type: ["string", "null"]
      required:
        - id
        - type
        - title
        - status
        - createdAt

    DashboardData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardItem"
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required:
        - items

    # ─── Partner ──────────────────────────────────────────────────────────────
    KpiMonth:
      type: object
      properties:
        month:
          type: string
        applications:
          type: integer
        rebatesCommittedAed:
          type: integer
        emiratiRegistered:
          type: integer
        productionsMatched:
          type: integer
        festivalSubmissions:
          type: integer
      required:
        - month
        - applications
        - rebatesCommittedAed
        - emiratiRegistered
        - productionsMatched
        - festivalSubmissions

    PartnerKpis:
      type: object
      properties:
        totals:
          type: object
          properties:
            applications:
              type: integer
            rebatesCommittedAed:
              type: integer
            emiratiRegistered:
              type: integer
            productionsMatched:
              type: integer
            festivalSubmissions:
              type: integer
          required:
            - applications
            - rebatesCommittedAed
            - emiratiRegistered
            - productionsMatched
            - festivalSubmissions
        monthly:
          type: array
          items:
            $ref: "#/components/schemas/KpiMonth"
      required:
        - totals
        - monthly