# dubaimedia/lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dubai Media Gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { UserTier } from './userTier';

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  tier: UserTier;
  analysisCount: number;
  createdAt: string;
}